requirements.txt
pyTelegramBotAPI
qrcode
python-dotenv